change lock in selectDevice action
